% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cv.R
\name{read_cv}
\alias{read_cv}
\title{Read controlled vocabularies (CV) in EAF/ETF file(s)}
\usage{
read_cv(file, full_path = FALSE, ecv = TRUE, progress = FALSE)
}
\arguments{
\item{file}{Path(s) to ELAN file(s) (.eaf/.etf/.ecv)}

\item{full_path}{Whether file names should include full path (defaults to `FALSE`)}

\item{ecv}{Whether to read linked external CVs as well (defaults to `TRUE`)}

\item{progress}{Whether to display a progress bar (defaults to `FALSE`)}
}
\value{
Data frame of the ELAN media metadata
}
\description{
Reads the controlled vocabularies in ELAN (.eaf/.etf/.ecv) file(s)
from a single file or a vector of files as input.
}
\details{
If `full_path` is set to `FALSE` (default), the input file names
will be listed with their base names only, without the full path.

If `progress` is set to `TRUE` (defaults to `FALSE`), a progress bar
is printed to the console to show the iteration across input files.
}
\examples{
eaf_file <- system.file("extdata", "example.eaf", package = "readelan", mustWork = TRUE)
read_cv(eaf_file, ecv = FALSE)

# External CVs (.ecv files) can also be read
ecv_file <- system.file("extdata", "syntax.ecv", package = "readelan", mustWork = TRUE)
read_cv(ecv_file, ecv = FALSE)

}
