% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eaf.R
\name{read_eaf}
\alias{read_eaf}
\title{Read annotations in EAF file(s)}
\usage{
read_eaf(
  file,
  tiers,
  xpath,
  fill_times = TRUE,
  full_path = FALSE,
  progress = FALSE
)
}
\arguments{
\item{file}{Path(s) to ELAN annotation file(s) (.eaf)}

\item{tiers}{Specify tiers (TIER_ID or TIER_TYPE) to be read}

\item{xpath}{Specify a detailed XPath for tiers to be read}

\item{fill_times}{Fill empty time slots of child annotations (defaults to `TRUE`)}

\item{full_path}{Whether file names should include full path (defaults to `FALSE`)}

\item{progress}{Whether to display a progress bar (defaults to `FALSE`)}
}
\value{
Data frame of ELAN annotations
}
\description{
Reads the annotations in ELAN annotation file(s) (.eaf)
from a single file or a vector of files as input.
}
\details{
Specific tiers or tier types can be targeted with a named list,
`tiers = list(tier = "tier1", tier_type = c("tier_type1", "tier_type2"))`,
or a custom XPath input, `xpath = ".//TIER[starts-with(@TIER_ID,'gloss')]"`.
A valid `tier` input will override any simultaneous `xpath` input.

If `fill_times` is set to `TRUE` (default), the time slots of child
annotations will be filled with the values of their parent annotations.

If `full_path` is set to `FALSE` (default), the input file names
will be listed with their base names only, without the full path.

If `progress` is set to `TRUE` (defaults to `FALSE`), a progress bar
is printed to the console to show the iteration across input files.
}
\examples{
eaf_file <- system.file("extdata", "example.eaf", package = "readelan", mustWork = TRUE)
read_eaf(eaf_file)

}
