% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_media.R
\name{read_media}
\alias{read_media}
\title{Read media metadata in EAF file(s)}
\usage{
read_media(file, full_path = FALSE, progress = FALSE)
}
\arguments{
\item{file}{Path(s) to ELAN annotation file(s) (.eaf)}

\item{full_path}{Whether file names should include full path (defaults to `FALSE`)}

\item{progress}{Whether to display a progress bar (defaults to `FALSE`)}
}
\value{
Data frame of ELAN media metadata
}
\description{
Reads the media metadata in ELAN (.eaf) file(s)
from a single file or a vector of files as input.
}
\details{
If `full_path` is set to `FALSE` (default), the input file names
will be listed with their base names only, without the full path.

If `progress` is set to `TRUE` (defaults to `FALSE`), a progress bar
is printed to the console to show the iteration across input files.
}
\examples{
eaf_file <- system.file("extdata", "example.eaf", package = "readelan", mustWork = TRUE)
read_media(eaf_file)

}
