% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tiers.R
\name{read_tiers}
\alias{read_tiers}
\title{Read tier metadata in in EAF/ETF/ECV file(s)}
\usage{
read_tiers(file, full_path = FALSE, progress = FALSE)
}
\arguments{
\item{file}{Path(s) to ELAN file(s) (.eaf/.etf/.ecv)}

\item{full_path}{Whether file names should include full path (defaults to `FALSE`)}

\item{progress}{Whether to display a progress bar (defaults to `FALSE`)}
}
\value{
Data frame of ELAN tier metadata
}
\description{
Reads the tier metadata in ELAN file(s) (.eaf/.etf/.ecv)
from a single file or a vector of files as input.
}
\details{
If `full_path` is set to `FALSE` (default), the input file names
will be listed with their base names only, without the full path.

If `progress` is set to `TRUE` (defaults to `FALSE`), a progress bar
is printed to the console to show the iteration across input files.
}
\examples{
eaf_file <- system.file("extdata", "example.eaf", package = "readelan", mustWork = TRUE)
read_tiers(eaf_file)

etf_file <- system.file("extdata", "CLP_annotation_v2-2.etf", package = "readelan", mustWork = TRUE)
read_tiers(etf_file)

}
